requirements.txt
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.5.23
pydantic-settings>=2.0.0,<3.0.0
openai==1.7.0
httpx>=0.25.0,<0.28.0
streamlit==1.30.0
pypdf==3.17.4
python-dotenv==1.0.0
tiktoken==0.5.2; python_version < "3.14"
sentence-transformers==2.2.2
transformers==4.36.2
tokenizers==0.15.2
huggingface_hub==0.19.4
rank-bm25==0.2.2
